+++ web/src/test/java/ua/com/vertex/dao/CourseDaoForTest.java
@@ -1,0 +1,51 @@
+package ua.com.vertex.dao;
+
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataAccessException;
+import org.springframework.jdbc.core.namedparam.MapSqlParameterSource;
+import org.springframework.jdbc.core.namedparam.NamedParameterJdbcTemplate;
+import org.springframework.jdbc.support.GeneratedKeyHolder;
+import org.springframework.jdbc.support.KeyHolder;
+import org.springframework.stereotype.Repository;
+import ua.com.vertex.beans.Course;
+
+
+import javax.sql.DataSource;
+
+@Repository
+public class CourseDaoForTest implements ua.com.vertex.dao.interfaces.CourseDaoForTest {
+
+    private final NamedParameterJdbcTemplate jdbcTemplate;
+
+    private static final String COLUMN_COURSE_NAME = "name";
+    private static final String COLUMN_COURSE_START = "start";
+    private static final String COLUMN_COURSE_FINISHED = "finished";
+    private static final String COLUMN_COURSE_PRICE = "price";
+    private static final String COLUMN_COURSE_TEACHER_NAME = "teacher_name";
+    private static final String COLUMN_COURSE_SCHEDULE = "schedule";
+    private static final String COLUMN_COURSE_NOTES = "notes";
+
+    @Override
+    public int createCourse(Course course) throws DataAccessException {
+        String query = "INSERT INTO Courses (name, start, finished, price, teacher_name, schedule, notes) " +
+                "VALUES(:name, :start, :finished, :price, :teacher_name, :schedule, :notes)";
+
+        KeyHolder keyHolder = new GeneratedKeyHolder();
+        MapSqlParameterSource source = new MapSqlParameterSource();
+        source.addValue(COLUMN_COURSE_NAME, course.getName());
+        source.addValue(COLUMN_COURSE_START, course.getStart());
+        source.addValue(COLUMN_COURSE_FINISHED, course.isFinished());
+        source.addValue(COLUMN_COURSE_PRICE, course.getPrice());
+        source.addValue(COLUMN_COURSE_TEACHER_NAME, course.getTeacherName());
+        source.addValue(COLUMN_COURSE_SCHEDULE, course.getSchedule());
+        source.addValue(COLUMN_COURSE_NOTES, course.getNotes());
+
+        jdbcTemplate.update(query, source, keyHolder);
+        return keyHolder.getKey().intValue();
+    }
+
+    @Autowired
+    public CourseDaoForTest(DataSource dataSource) {
+        this.jdbcTemplate = new NamedParameterJdbcTemplate(dataSource);
+    }
+}
